+++ app/src/main/java/com/tech/mymovietvshows/Model/MovieCreditsCrewModel.java
@@ -4,7 +4,7 @@ public class MovieCreditsCrewModel {
 
     private boolean adult;
     private int gender;
-    private long id;
+    private int id;
     private String known_for_department;
     private String name;
     private String original_name;
@@ -17,7 +17,7 @@ public class MovieCreditsCrewModel {
     public MovieCreditsCrewModel() {
     }
 
-    public MovieCreditsCrewModel(boolean adult, int gender, long id, String known_for_department, String name, String original_name, Double popularity, String profile_path, String credit_id, String department, String job) {
+    public MovieCreditsCrewModel(boolean adult, int gender, int id, String known_for_department, String name, String original_name, Double popularity, String profile_path, String credit_id, String department, String job) {
         this.adult = adult;
         this.gender = gender;
         this.id = id;
@@ -47,11 +47,11 @@ public class MovieCreditsCrewModel {
         this.gender = gender;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
